# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API documentation for .",
        "title": "Rebellion Banking APII",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/sms-agent/v1/messages": {
            "get": {
                "security": [
                    {
                        "SMSAgentKey": []
                    }
                ],
                "description": "Returns queued SMS messages for the Android agent to deliver locally via Shizuku.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms agent"
                ],
                "summary": "Poll pending SMS for mobile agent",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Max messages",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/internal_modules_smsagent_transport_http.MessageResponse"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    }
                }
            }
        },
        "/sms-agent/v1/messages/{id}/ack": {
            "post": {
                "security": [
                    {
                        "SMSAgentKey": []
                    }
                ],
                "description": "Marks a queued SMS as delivered or failed after the agent attempts local injection.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "sms agent"
                ],
                "summary": "Acknowledge SMS delivery",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Message ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Delivery result",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AckMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_smsagent_transport_http.AgentErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "internal_modules_smsagent_transport_http.AckMessageRequest": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "deviceId": {
                    "type": "string",
                    "example": "abc123"
                },
                "errorMessage": {
                    "type": "string"
                },
                "status": {
                    "type": "string",
                    "example": "delivered"
                }
            }
        },
        "internal_modules_smsagent_transport_http.AgentErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_modules_smsagent_transport_http.MessageResponse": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bank": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        }
    }
}